Add user input for borrower and form verification.
This commit adds multiple things from the past week. Changes include
input for adding the name of a borrower, as well as form verification to
assure that a name has been inputted. Changes also include fixes to the
book descriptions, and consolidation of the verification controller into
the book controller.