Add ASP.Net Web API Files, add project files to solution, move old project files to old directory.
In this commit, I added a new Visual Studio ASP.Net Web API project to
the project directory. Inside is a ton of new files, as well as the
original project files which I have now added to the Visual Studio
project solution. The old files have been moved to a separate directory
in case something messes up.