Add book selection functionality.
This commit adds book selection functionality, where the user can now
select the book for which to view details, including the borrowing
functionality. This commit though does not contain the fully working
version of this functionality. While the user can select a book, the
program does not yet display book details.