Finish book selection functionality.
This commit finishes the book selection functionality by improving how
the Web API serves data. Instead of serving it all as a string, it
parses the JSON server-side and then sends that over, and vice-versa
when it receives an array of JS objects. This also allows for future
adding of books through the client.